Move catch piece files
